refactor(index): widen FtsHit.score to f64 to preserve BM25 precision
Remove the gratuitous `as f32` narrowing of the raw f64 bm25() value.
The score field and the negation expression both become f64; the fusion
pipeline only consumes the id so no downstream type change is needed.

+++ src/index/fts.rs
@@ -96,10 +96,7 @@ impl Fts {
         let rows = match stmt.query_map(rusqlite::params![query, limit as i64], |row| {
             let id: String = row.get(0)?;
             let raw: f64 = row.get(1)?;
-            Ok(FtsHit {
-                id,
-                score: -raw as f32,
-            })
+            Ok(FtsHit { id, score: -raw })
         }) {
             Ok(r) => r,
             Err(e) => {
@@ -145,5 +142,5 @@ pub struct FtsHit {
     /// Memory id stored in the `UNINDEXED` payload column.
     pub id: String,
     /// `-bm25()` so higher is more relevant.
-    pub score: f32,
+    pub score: f64,
 }
